add welding subfolder
