+++ bam_masterdata/datamodel/welding/__init__.py
@@ -1,0 +1,1 @@
+"""Submodule containing type definitions related to welding experiments."""
